+++ Cargo.toml
@@ -49,8 +49,6 @@ clap = { version = "4.3.1", features = ["derive", "env"] }
 tracing = "0.1.37"
 tracing-subscriber = { version = "0.3.17", features = ["json", "env-filter", "std"]}
 pg_query = "0.7"
-hex = "0.4"
-humantime = "2.1"
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
 jemallocator = "0.5.0"
